Rename to IdentityResponse, remove already impl DidDocument
